[REFACTOR] Rename auth reducer resources to ensure clarity and differentiation
