+++ src/lib/store/store.ts
@@ -1,13 +1,12 @@
 import { configureStore } from '@reduxjs/toolkit';
 
+import authReducer from '@/modules/auth/reducers/authReducer';
 import postsReducer from '@/modules/posts/reducers';
 import storiesReducer from '@/modules/stories/reducers';
 
-import authReducer from './reducers/authReducer';
-
 export const store = configureStore({
   reducer: {
-    authReducer,
+    auth: authReducer,
     posts: postsReducer,
     stories: storiesReducer,
   },
